--- src/main/java/org/scoula/member/dto/LoginResponseDTO.java
@@ -1,11 +1,0 @@
-package org.scoula.member.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class LoginResponseDTO {
-    private String token;
-    private String username;
-}
