--- src/main/java/org/scoula/member/dto/PasswordResetDTO.java
@@ -1,10 +1,0 @@
-package org.scoula.member.dto;
-
-import lombok.Data;
-
-@Data
-public class PasswordResetDTO {
-    private String loginId;
-    private String password;
-}
-
