+++ src/main/java/org/scoula/member/service/MemberServiceImpl.java
@@ -51,6 +51,9 @@ public class MemberServiceImpl implements MemberService {
     }
   }
 
+  public MemberVO findByUsername(String loginId) {
+    return mapper.findByUsername(loginId);
+  }
   // 회원 가입
   @Transactional
   @Override
@@ -73,35 +76,7 @@ public class MemberServiceImpl implements MemberService {
     return get(member.getLoginId());
   }
 
-  @Override
-  public Optional<MemberDTO> login(String username, String password) {
-    MemberVO member = mapper.get(username);
 
-    if (member != null && passwordEncoder.matches(password, member.getPassword())) {
-      return Optional.of(MemberDTO.of(member));
-    }
-
-    return Optional.empty();  // 비밀번호 불일치 또는 사용자 없음
-  }
-
-  public MemberVO findByUsername(String loginId) {
-    return mapper.findByUsername(loginId);
-  }
-
-  @Override
-  public boolean resetPassword(String loginId, String password) {
-    MemberVO member = mapper.get(loginId);
-    if (member == null) return false;
-
-    String encrypted = passwordEncoder.encode(password);
-    mapper.resetPassword(loginId, encrypted);
-    return true;
-  }
-
-  @Override
-  public MemberVO findByEmail(String email) {
-    return mapper.findByEmail(email);
-  }
 
 
 }
