+++ src/main/java/org/scoula/security/dto/PasswordResetDTO.java
@@ -1,9 +1,10 @@
-package org.scoula.member.dto;
+package org.scoula.security.dto;
 
 import lombok.Data;
 
 @Data
-public class LoginRequestDTO {
+public class PasswordResetDTO {
     private String loginId;
     private String password;
 }
+
